the components splited as server and clientside
