+++ app/components/ClientComponet/LastMinutePackage.tsx
@@ -4,8 +4,8 @@ import React from "react";
 import Image from "next/image";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faStar, faCalendarAlt } from "@fortawesome/free-solid-svg-icons";
-import img from "../../public/TourPackageimg.jpg";
-import CardDetails from "./CardDetails";
+import img from "../../../public/TourPackageimg.jpg";
+import CardDetails from "../ServerComponents.tsx/CardDetails";
 import { useState, useEffect } from "react";
 
 interface UserData {
