+++ app/components/ClientComponet/TourPackage.tsx
@@ -1,7 +1,7 @@
 "use client";
 import React from "react";
 import Image from "next/image";
-import img from "../../public/TourPackageimg.jpg";
+import img from "../../../public/TourPackageimg.jpg";
 import { useState, useEffect } from "react";
 import { faCalendarAlt } from "@fortawesome/free-regular-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
