+++ app/components/ServerComponents.tsx/CardDetails.tsx
@@ -1,8 +1,6 @@
-"use client";
 import React from "react";
-import Image from "next/image";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faStar, faCalendarAlt } from "@fortawesome/free-solid-svg-icons";
+import { faStar } from "@fortawesome/free-solid-svg-icons";
 
 interface UserData {
   id: number;
