+++ app/components/ServerComponents.tsx/LastMinuteDeals.tsx
